seturity/requirements.txt
aiohttp>=3.9.0
web3>=6.0.0
redis>=5.0.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
